windows-client/requirements.txt
pynput
websockets
zeroconf
pystray
Pillow
netifaces
